pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "django-shortcuts==1.6",
    "django==5.1.3",
    "flask==3.1.0",
    "requests>=2.32.3",
    "snowflake-connector-python==3.12.3",
]
